Adds default config for assets pallet
